airflow/requirements.txt
requests
beautifulsoup4==4.11.1
redis==4.3.4
kafka-python==2.0.2
atoma==0.0.17